encapsulate CFlattenRegionsStreamHashImageFilter.h's logic into ICPUImage.cpp & override `core::blake3_hash_t ICPUImage::computeContentHash() const` with it, leave some comments for Matt & remove old CMake comment from src/nbl/CMakeLists.txt, update examples_tests submodule
