+++ include/nbl/asset/ICPUImage.h
@@ -14,7 +14,7 @@
 namespace nbl::asset
 {
 
-class ICPUImage final : public IImage, public IPreHashed
+class NBL_API2 ICPUImage final : public IImage, public IPreHashed
 {
 	public:
 		inline static core::smart_refctd_ptr<ICPUImage> create(const SCreationParams& _params)
@@ -47,41 +47,7 @@ class ICPUImage final : public IImage, public IPreHashed
 		// Having regions specififed to upload is optional!
 		inline size_t getDependantCount() const override {return !missingContent()&&buffer ? 1:0;}
 
-		//!
-		inline core::blake3_hash_t computeContentHash() const override
-		{
-			// TODO: Arek turn this into an image filter maybe?
-			::blake3_hasher hasher;
-			::blake3_hasher_init(&hasher);
-			for (auto m=0; m<m_creationParams.mipLevels; m++)
-			{
-				const auto blockInfo = getTexelBlockInfo();
-				const auto mipExtentInBlocks = blockInfo.convertTexelsToBlocks(getMipSize(m));
-				::blake3_hasher levelHasher;
-				::blake3_hasher_init(&levelHasher);
-				if (missingContent())
-				{
-					const auto zeroLength = blockInfo.getBlockByteSize()*mipExtentInBlocks[0];
-					auto zeroArray = std::make_unique<uint8_t[]>(zeroLength);
-					for (auto l=0; l<m_creationParams.arrayLayers; l++)
-					{
-						// layers could be run in parallel
-						::blake3_hasher layerHasher;
-						::blake3_hasher_init(&layerHasher);
-						for (auto z=0; z<mipExtentInBlocks[2]; z++)
-						for (auto y=0; y<mipExtentInBlocks[1]; y++)
-							::blake3_hasher_update(&layerHasher,zeroArray.get(),zeroLength);
-						core::blake3_hasher_update(levelHasher,core::blake3_hasher_finalize(layerHasher));
-					}
-				}
-				else
-				{
-					_NBL_TODO(); // TODO: Arek
-				}
-				core::blake3_hasher_update(hasher,core::blake3_hasher_finalize(levelHasher));
-			}
-			return core::blake3_hasher_finalize(hasher);
-		}
+		core::blake3_hash_t computeContentHash() const override;
 
 		inline bool missingContent() const override
 		{
